v1.1.0-release

+++ System/DataArchiveOperation/SQLDataArchiveOperation.cs
@@ -6,16 +6,23 @@
 //-------------------------------------------------
 
 using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Data;
 using System.IO;
+using System.Text;
 using System.Threading.Tasks;
+using FlexiArchiveSystem.ArchiveOperation.IO;
+using FlexiArchiveSystem.Assist;
+using FlexiArchiveSystem.Setting;
 using Mono.Data.Sqlite;
 
 namespace FlexiArchiveSystem.ArchiveOperation
 {
-    public class SQLDataArchiveOperation : IDataArchiveOperation, ISetDataArchivePath, ICloneDataArchive
+    internal class SQLDataArchiveOperation : IDataArchiveOperation, ISetDataArchivePath, ICloneDataArchive
     {
+        private string _ArchiveSystemName;
+        public string ArchiveSystemName => _ArchiveSystemName;
         public bool IsValidation
         {
             get { return IsActive && Directory.Exists(Path); }
@@ -57,12 +64,14 @@ namespace FlexiArchiveSystem.ArchiveOperation
 
         public void SetDataArchiveOperationHelper(DataArchiveOperationHelper helper)
         {
-            helper.SetArchiveID(_archiveID);
+            helper.Init(ArchiveSystemName);
+            helper.UpdateDirtyState(_archiveID);
             archiveOperationHelper = helper;
         }
 
-        public void Init(int archiveID)
+        public void Init(string moudleName,int archiveID)
         {
+            _ArchiveSystemName = moudleName;
             SetArchiveID(archiveID);
             dataMap = new Dictionary<string, Dictionary<string, string>>();
             FilePath = GetAndCombineDataFilePath();
@@ -96,16 +105,13 @@ namespace FlexiArchiveSystem.ArchiveOperation
             connection.Open();
         }
 
-        public void DataPersistent(string key, string dataStr)
+        public void DataPersistent(string groupKey, string dataKey, string dataStr)
         {
             if (connection == null)
             {
                 InitDBConnection();
             }
-
-            var keyTuple = DataKeyHandler.GetAndProcessKeyCollection(key);
-            string groupKey = keyTuple.Item1;
-            string dataKey = keyTuple.Item2;
+            
 
             string queryGroupTableCmd = $"select name from sqlite_master where type='table' and name='{groupKey}'";
             SqliteCommand groupCommand = new SqliteCommand(queryGroupTableCmd, connection);
@@ -136,11 +142,80 @@ namespace FlexiArchiveSystem.ArchiveOperation
             bool hasExistedBefore = string.IsNullOrEmpty(lastResult) == false;
             if (hasExistedBefore == false)
             {
-                TryRecordKey(groupKey, dataKey);
+                TryRecordKey(groupKey);
             }
         }
 
-        public string Read(string key)
+        public void DataPersistent(params DataObject[] dataObjects)
+        {
+            foreach (var dataObject in dataObjects)
+            {
+                var keyTuple = DataKeyHandler.GetAndProcessKeyCollection(dataObject.Key);
+                string groupKey = keyTuple.Item1;
+                string dataKey = keyTuple.Item2;
+                DataPersistent(groupKey, dataKey, dataObject.GetSerializeData());
+            }
+        }
+
+        public async Task DataPersistentAsync(string groupKey, string dataKey, string dataStr, Action complete)
+        {
+            if (connection == null)
+            {
+                InitDBConnection();
+            }
+
+            string queryGroupTableCmd = $"select name from sqlite_master where type='table' and name='{groupKey}'";
+            SqliteCommand groupCommand = new SqliteCommand(queryGroupTableCmd, connection);
+            var cmdReader = await (groupCommand.ExecuteReaderAsync());
+            bool exsitTable = (await cmdReader.ReadAsync());
+            
+            if (exsitTable == false)
+            {
+                //create table
+                string createTableCmd = $"create table {groupKey}(dataKey varchar(32) PRIMARY KEY,data varchar(32));";
+                SqliteCommand createTableCommand = new SqliteCommand(createTableCmd, connection);
+                await createTableCommand.ExecuteNonQueryAsync();
+            }
+
+            string updateDataCmd = $"INSERT OR REPLACE into {groupKey} (dataKey,data) values ('{dataKey}','{dataStr}')";
+            SqliteCommand queryDataCommand = new SqliteCommand(updateDataCmd, connection);
+            await queryDataCommand.ExecuteNonQueryAsync();
+
+            bool isExist = TryGetData(groupKey, dataKey, out string lastResult);
+            if (isExist == false)
+            {
+                if (dataMap.ContainsKey(groupKey) == false)
+                {
+                    dataMap.Add(groupKey, new Dictionary<string, string>());
+                }
+            }
+
+            dataMap[groupKey][dataKey] = dataStr;
+            bool hasExistedBefore = string.IsNullOrEmpty(lastResult) == false;
+            if (hasExistedBefore == false)
+            {
+                TryRecordKey(groupKey);
+            }
+            
+            complete?.Invoke();
+        }
+
+        public async Task DataPersistentAsync(Action complete, params DataObject[] dataObjects)
+        {
+            IList<Task> writeTasks = new List<Task>();
+            foreach (var dataObject in dataObjects)
+            {
+                var keyTuple = DataKeyHandler.GetAndProcessKeyCollection(dataObject.Key);
+                string groupKey = keyTuple.Item1;
+                string dataKey = keyTuple.Item2;
+                writeTasks.Add( DataPersistentAsync(groupKey, dataKey, dataObject.GetSerializeData(),null));
+            }
+
+            await Task.WhenAll(writeTasks);
+            complete?.Invoke();
+        }
+
+        public string Read(string groupKey, string dataKey)
         {
             if (File.Exists(FilePath) == false)
             {
@@ -151,10 +226,6 @@ namespace FlexiArchiveSystem.ArchiveOperation
             {
                 InitDBConnection();
             }
-
-            var keyTuple = DataKeyHandler.GetAndProcessKeyCollection(key);
-            string groupKey = keyTuple.Item1;
-            string dataKey = keyTuple.Item2;
 
             bool isExist = TryGetData(groupKey, dataKey, out string result);
             if (isExist)
@@ -200,7 +271,7 @@ namespace FlexiArchiveSystem.ArchiveOperation
             {
                 return;
             }
-
+            
             if (connection != null)
             {
                 try
@@ -210,20 +281,27 @@ namespace FlexiArchiveSystem.ArchiveOperation
                     SqliteConnection.ClearPool(connection);
                     connection = null;
                 }
-                catch (Exception)
+                catch (Exception e)
                 {
+                    throw e;
                 }
             }
 
             if (Directory.Exists(Path))
             {
+                bool isUse = await IOHelper.FileIsInUse(FilePath, 1000, 200);
+                if (isUse)
+                {
+                    Logger.LOG_ERROR($"文件{FilePath}长时间被占用，无法删除存档");
+                    return;
+                }
                 Directory.Delete(Path, true);
             }
 
             this.Dispose();
         }
 
-        public void Delete(string key)
+        public void Delete(string groupKey, string dataKey)
         {
             if (File.Exists(FilePath) == false)
             {
@@ -235,14 +313,16 @@ namespace FlexiArchiveSystem.ArchiveOperation
                 InitDBConnection();
             }
 
-            var keyTuple = DataKeyHandler.GetAndProcessKeyCollection(key);
-            string groupKey = keyTuple.Item1;
-            string dataKey = keyTuple.Item2;
-
             string delDataCmd = $"delete from {groupKey} where dataKey='{dataKey}'";
             SqliteCommand delDataCommand = new SqliteCommand(delDataCmd, connection);
             delDataCommand.ExecuteNonQuery();
-            dataMap.Remove(key);
+            if (dataMap != null)
+            {
+                if (dataMap.TryGetValue(groupKey, out var value))
+                {
+                    value.Remove(dataKey);
+                }
+            }
         }
 
         public void DeleteGroup(string groupKey)
@@ -267,29 +347,40 @@ namespace FlexiArchiveSystem.ArchiveOperation
             }
         }
 
-        public void TryRecordKey(string groupKey, string dataKey)
+        public void TryRecordKey(string groupKey)
         {
             if (AllGroupKeys != null)
             {
                 AllGroupKeys.Add(groupKey);
             }
 
-            ArchiveOperationHelper.RecordKey(_archiveID, groupKey, dataKey);
+            ArchiveOperationHelper.RecordKey(_archiveID, groupKey);
         }
 
         public void TryRemoveAllGroupKey()
         {
             ArchiveOperationHelper.DeleteAllGroupKeyFromDisk();
         }
-
-        public IDataArchiveSourceWrapper GetSource()
+        
+        private async Task<List<string>> LoadAllGroupKeyFromDisk()
         {
+            return await ArchiveOperationHelper.GetAllGroupKey();
+        }
+
+#pragma warning disable CS1998
+        public async Task<IDataArchiveSourceWrapper> GetSource()
+        {
+            if (AllGroupKeys == null)
+            {
+                AllGroupKeys = await LoadAllGroupKeyFromDisk();
+            }
             SqliteArchiveSourceWrapper wrapper = new SqliteArchiveSourceWrapper();
             wrapper.sourcePath = FilePath;
             return wrapper;
         }
-
-        public void CloneTo(IDataArchiveSourceWrapper source)
+#pragma warning restore CS1998
+        
+        public async Task CloneTo(IDataArchiveSourceWrapper source)
         {
             var wrapper = source as SqliteArchiveSourceWrapper;
 
@@ -302,27 +393,63 @@ namespace FlexiArchiveSystem.ArchiveOperation
             {
                 Directory.CreateDirectory(Path);
             }
+            
+            byte[] buffer = new byte[0x1000];
+            int readLen;
+            bool isUse = await IOHelper.FileIsInUse(wrapper.sourcePath, 1000 ,200);
+            if (isUse)
+            {
+                Logger.LOG_ERROR($"文件{wrapper.sourcePath}长时间被占用，无法克隆存档");
+                return;
+            }
 
-            File.Copy(wrapper.sourcePath, FilePath);
+            using var reader = new FileStream(wrapper.sourcePath, FileMode.Open, 
+                FileAccess.Read, FileShare.None, bufferSize: 4096, useAsync: true);
+            using var writer = new FileStream(FilePath, FileMode.OpenOrCreate, 
+                FileAccess.Write, FileShare.None, bufferSize: 4096, useAsync: true);
+            while ((readLen = await reader.ReadAsync(buffer, 0, buffer.Length)) != 0)
+            {
+                await writer.WriteAsync(buffer, 0, buffer.Length);
+            }
+            ArchiveOperationHelper.RecordAllGroupKeyWhenClone(_archiveID);
+        }
+
+        public async Task CloseIOAsync()
+        {
+            if (connection != null)
+            {
+                await connection.DisposeAsync();
+            }
         }
 
         public void Dispose()
         {
             if (connection != null)
             {
-                try
-                {
-                    // connection.Dispose();
-                    SqliteConnection.ClearPool(connection);
-                }
-                catch (Exception)
-                {
-                }
+                connection.Dispose();
             }
 
+            archiveOperationHelper = null;
             connection = null;
             IsActive = false;
+            GC.Collect();
+            GC.WaitForPendingFinalizers();
         }
+
+        public async Task DisposeAsync()
+        {
+            if (connection != null)
+            {
+                await connection.DisposeAsync();
+            }
+
+            archiveOperationHelper = null;
+            connection = null;
+            IsActive = false;
+            GC.Collect();
+            GC.WaitForPendingFinalizers();
+        }
+        
 
         private string GetAndCombineDataFilePath()
         {
